requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
aniso8601==9.0.1
anyio==4.12.0
beautifulsoup4==4.9.3
CacheControl==0.14.3
cachetools==6.2.3
certifi==2022.6.15
cffi==2.0.0
chardet==4.0.0
click==8.1.3
colorama==0.4.5
cryptography==46.0.3
exceptiongroup==1.3.1
firebase_admin==7.1.0
Flask==2.1.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.7.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==1.1.3
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==19.6.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==2.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.1.2
numpy==1.24.2
opencv-python-headless==4.7.0.68
Pillow==9.4.0
proto-plus==1.26.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2022.2.1
requests==2.25.1
rsa==4.9.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
typing_extensions==4.15.0
urllib3==1.26.12
Werkzeug==2.0.3
zipp==3.8.1
dependency-injector==4.41.0
